streva_backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
